perf: 主机与主机关系事件处理优化 #1145
主机缓存更新策略优化；
主机事件监听代码重构。

+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/impl/sync/HostRelationWatchThread.java
@@ -176,7 +176,7 @@ public class HostRelationWatchThread extends Thread {
      */
     private void updateTopoToHost(HostTopoDTO hostTopoDTO) {
         // 若主机存在需将拓扑信息同步至主机信息冗余字段
-        long affectedNum = applicationHostDAO.syncHostTopo(dslContext, hostTopoDTO.getHostId());
+        int affectedNum = applicationHostDAO.syncHostTopo(dslContext, hostTopoDTO.getHostId());
         if (affectedNum == 0) {
             log.info("no host topo synced");
         } else if (affectedNum < 0) {
